requirements.txt
cx_Oracle
mysql-connector-python
pandas
